refactor Node into its own file
